Add ALT+X keybind to open xkill:
This is a hack to prevent my computer from locking when I don't want it to.
Slock fails to lock the screen if a program is attached to the cursor (don't
know the proper x11 term for this), and xkill happens to grab the cursor when
you start it, preventing slock from locking. The DPMS system still kicks in on
my monitor after 5 minutes though which is annoying, especially since it takes
2-3 seconds for it to come back on after hitting a key.

Suffice to say, I should probably find a more elegant way of handling this.